Derive channel name from bridge info if no canonical alias exists
The channel name will be "name:network.protocol", or "name:protocol" if
no network info exists in the bridge. Channels with names that don't
start with a CHANPREFIX are prefixed with @ on the assumption that they
are DM pseudochannels, which produces the correct behaviour for Discord
and Googlechat bridges (but not for Matrix and possibly not for other
bridges, either).